archive old pins

+++ commands/learn/forget_command.rb
@@ -16,7 +16,9 @@ module Commands
         }
       end
 
-      def command(_, name)
+      def command(e, name)
+        return if e.from_bot?
+
         command = UserCommand.find_by(name: name)
         if command
           command.destroy
